fixed the loop in the client so now works ok
fixed the loop in the client so now works OK.. going to abandon joining the telemetry with the control, getting messy. will create inside the client for lights a senor that broadcast name and if on or off. i''ll bing that into \lights data then with it's own gRPC service